chore(audit): full-stack alignment fixes — wayback save 404, skin nav parity, doc counts
Fixes a cluster of audit findings around web/cli/api alignment, skin parity,
and doc/code drift. No breaking changes; all routes/params preserved.

Bug fixes
- panel: Wayback Save 路径错位 — `/api/v1/wayback/save` → `/api/v1/admin/wayback/save`
  (admin_router 的 prefix 是 /api/v1/admin；前端原路径在所有皮肤都 404)
- panel/souwen-google: PAGE_TITLE_KEYS 补 `/warp`，避免标题 fallback 到 dashboard
- panel/{carbon,apple,ios}: NAV_ITEMS 补 `/video`、`/tools` 入口（routes.tsx 已注册但侧栏不可达）

Refactor
- server/routes/admin: 把 `/admin/wayback/save` 从 admin/warp.py 拆出到独立的
  admin/wayback.py（SRP）；admin_router 在 __init__ 中聚合新模块。功能/路径不变。

Docs
- README / README.en: fetch 横切能力计数 15+4 → 16+4 = 20 源
- api-reference: 示例 version 0.6.3 → 1.1.1 + "动态返回，示例值仅作参考"
- api-reference: fetch provider 列表补 `arxiv_fulltext`，"19 提供者" → "20 提供者"
- panel/{souwen-google,souwen-nebula}: MainLayout NAV_ITEMS 注释更新到当前 9 项
- CHANGELOG: 顶部新增 [Unreleased] — 全面评审修复 段落

Tests
- pytest tests/ → 856 passed + 5 skipped
- ruff check / format → clean
- python -c "from souwen.server.app import app" → 49 routes 全解析正常
