+++ panel/src/skins/apple/components/layout/MainLayout.tsx
@@ -39,6 +39,8 @@ import {
   Layers,
   Check,
   FileText,
+  Play,
+  Wrench,
 } from 'lucide-react'
 import { useAuthStore } from '@core/stores/authStore'
 import { canAccessNavItem } from '@core/lib/access'
@@ -54,6 +56,8 @@ const NAV_ITEMS = [
   { to: '/', icon: LayoutDashboard, labelKey: 'nav.dashboard' },
   { to: '/search', icon: Search, labelKey: 'nav.search' },
   { to: '/fetch', icon: FileText, labelKey: 'nav.fetch' },
+  { to: '/video', icon: Play, labelKey: 'nav.video' },
+  { to: '/tools', icon: Wrench, labelKey: 'nav.tools' },
   { to: '/sources', icon: Database, labelKey: 'nav.sources' },
   { to: '/network', icon: Wifi, labelKey: 'nav.network' },
   { to: '/warp', icon: Shield, labelKey: 'nav.warp' },
